PalettedBlockArray: allow getting and providing palette as raw bytes, instead of an array
closes #36

This is significantly faster, particularly in the 16 bpb case, which
shows a combined performance improvement of 24500 ns -> 1700 ns for
recreating a paletted array using a string instead of an array.

We really were paying a hilariously large performance penalty for
processing these as an array, even just within the extension itself.
Not to mention the cost of needing to use pack() and unpack() in
"Fast"ChunkSerializer...