Revert "PalettedBlockArray: allow getting and providing palette as raw bytes, instead of an array" and related commits
It turns out this is only useful if we want to preserve the bytes for
direct reuse later. We don't really want to do this as discussed in
pmmp/PocketMine-MP#6786. When the data must be decomposed into integers,
transformed and repacked to binary, getting the data as binary first is
actually worse for performance (assuming that ByteBufferReader is used).
So it looks like we're stuck with arrays, for better or worse :(

This reverts commit 8e83c8d1d648e11532cbdcb59cbdaa5d66a7d08f.
This reverts commit 5614724b8da71d10603205bcd60e6752affcdd48.
This reverts commit a26bed68197691373cf43b35f4519f6b71f832b3.